requirements.txt
django
djangorestframework
requests
proxmoxer